♻️ Refactor: Appiontment 페이지와 주소 로직 분리 #43
